Move theme style to core
