+++ views/blogs/components/post/__tests__/Container.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import Container from '~/views/blogs/components/post/Container'
-import theme from '~/theme/blogTheme'
+import theme from '~/views/core/theme/styleguide/blogTheme'
 
 describe('Blog Container', () => {
   it('should render without throwing an error', function () {
