+++ views/blogs/components/post/__tests__/Content.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import Content from '~/views/blogs/components/post/Content'
-import theme from '~/theme/blogTheme'
+import theme from '~/views/core/theme/styleguide/blogTheme'
 
 describe('Blog Content', () => {
   it('should render without throwing an error', function () {
