+++ views/blogs/components/post/__tests__/PostInfo.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import PostInfo from '~/views/blogs/components/post/PostInfo'
-import theme from '~/theme/blogTheme'
+import theme from '~/views/core/theme/styleguide/blogTheme'
 
 describe('Blog PostInfo', () => {
   it('should render without throwing an error', function () {
