+++ views/blogs/components/post/postinfo/__tests__/Author.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import Author from '../Author'
-import theme from '~/theme/blogTheme'
+import theme from '~/views/core/theme/styleguide/blogTheme'
 
 describe('Blog Post Info Author', () => {
   it('should render without throwing an error', function () {
