+++ views/blogs/components/post/postinfo/__tests__/Container.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import Container from '../Container'
-import theme from '~/theme/blogTheme'
+import theme from '~/views/core/theme/styleguide/blogTheme'
 
 describe('Blog Post Info Container', () => {
   it('should render without throwing an error', function () {
