+++ views/blogs/components/post/postinfo/__tests__/SubInfo.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallowWithBaseTheme } from '~/views/core/utils/withThemeProviders'
 import SubInfo from '../SubInfo'
-import theme from '~/theme/blogTheme'
+import theme from '~/views/core/theme/styleguide/blogTheme'
 
 describe('Blog Post Info SubInfo', () => {
   it('should render without throwing an error', function () {
